requirements.txt
PyPDF2==1.26.0
Pillow==8.2.0
